@@ -22,8 +22,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-"use strict";
-
 class Failure extends Error {
   constructor(message, code = null){
     super(message);
@@ -38,4 +36,13 @@ class Failure extends Error {
   }
 }
 
-module.exports = { Failure };
+const ErrorCode = [
+  "COM library for the calling thread already initialized with different threading model. Please use 'COINIT_MULTITHREADED'",
+  "Failed to initialize COM library for the calling thread",
+  "Failed to initialize security",
+  "Failed to create IWbemLocator object",
+  "Could not connect to ROOT\\CIMV2 WMI namespace",
+  "Could not set proxy blanket"
+];
+
+export { Failure, ErrorCode };
